chore(tests): group middleware testing to one module
