+++ tests/functional/event_handler/test_api_gateway.py
@@ -1749,132 +1749,3 @@ def test_route_match_prioritize_full_match():
     # THEN the static_handler should have been called, because it fully matches the path directly
     response_body = json.loads(response["body"])
     assert response_body["hello"] == "static"
-
-
-def test_route_with_middleware():
-    # GIVEN a Rest API Gateway proxy type event
-    app = ApiGatewayResolver(proxy_type=ProxyEventType.APIGatewayProxyEvent)
-
-    # define custom middleware to inject new argument - "custom"
-    def middleware_1(app, get_response, **kwargs):
-        # inject a variable into the kwargs
-        response = get_response(app, custom="custom", **kwargs)
-
-        return response
-
-    # define custom middleware to inject new argument - "another_one"
-    def middleware_2(app, get_response, **kwargs):
-        # inject a variable into the kwargs
-        response = get_response(app, another_one=6, **kwargs)
-
-        return response
-
-    @app.get("/my/path", middlewares=[middleware_1, middleware_2])
-    def get_lambda(another_one: int, custom: str) -> Response:
-        assert isinstance(app.current_event, APIGatewayProxyEvent)
-        assert another_one == 6
-        assert custom == "custom"
-        return Response(200, content_types.TEXT_HTML, "foo")
-
-    # WHEN calling the event handler
-    result = app(LOAD_GW_EVENT, {})
-
-    # THEN process event correctly
-    # AND set the current_event type as APIGatewayProxyEvent
-    assert result["statusCode"] == 200
-    assert result["multiValueHeaders"]["Content-Type"] == [content_types.TEXT_HTML]
-    assert result["body"] == "foo"
-
-
-def test_with_router_middleware():
-    # GIVEN a Rest API Gateway proxy type event
-    app = ApiGatewayResolver(proxy_type=ProxyEventType.APIGatewayProxyEvent)
-
-    # define custom middleware to inject new argument - "custom"
-    def middleware_1(app, get_response, **kwargs):
-        # inject a variable into the kwargs
-        response = get_response(app, custom="custom", **kwargs)
-
-        return response
-
-    # define custom middleware to inject new argument - "another_one"
-    def middleware_2(app, get_response, **kwargs):
-        # inject a variable into the kwargs
-        response = get_response(app, another_one=6, **kwargs)
-
-        return response
-
-    app.use([middleware_1, middleware_2])
-
-    @app.get("/my/path")
-    def get_lambda(another_one: int, custom: str) -> Response:
-        assert isinstance(app.current_event, APIGatewayProxyEvent)
-        assert another_one == 6
-        assert custom == "custom"
-        return Response(200, content_types.TEXT_HTML, "foo")
-
-    # WHEN calling the event handler
-    result = app(LOAD_GW_EVENT, {})
-
-    # THEN process event correctly
-    # AND set the current_event type as APIGatewayProxyEvent
-    assert result["statusCode"] == 200
-    assert result["multiValueHeaders"]["Content-Type"] == [content_types.TEXT_HTML]
-    assert result["body"] == "foo"
-
-
-def test_dynamic_route_with_middleware():
-    # GIVEN
-    app = ApiGatewayResolver()
-
-    def middleware_one(app, get_response, **kwargs):
-        # inject a variable into the kwargs
-        response = get_response(app, injected="injected_value", **kwargs)
-
-        return response
-
-    @app.get("/<name>/<my_id>", middlewares=[middleware_one])
-    def get_lambda(my_id: str, name: str, injected: str) -> Response:
-        assert name == "my"
-        assert injected == "injected_value"
-
-        return Response(200, content_types.TEXT_HTML, my_id)
-
-    # WHEN calling the event handler
-    result = app(LOAD_GW_EVENT, {})
-
-    # THEN
-    assert result["statusCode"] == 200
-    assert result["multiValueHeaders"]["Content-Type"] == [content_types.TEXT_HTML]
-    assert result["body"] == "path"
-
-
-def test_middleware_early_return():
-    # GIVEN
-    app = ApiGatewayResolver()
-
-    def middleware_one(app, get_response, **context):
-        # inject a variable into the kwargs
-        response = get_response(app, injected="injected_value", **context)
-
-        return response
-
-    def early_return_middleware(app, get_response, **context):
-        assert context.get("injected") == "injected_value"
-
-        return Response(400, content_types.TEXT_HTML, "bad_response")
-
-    @app.get("/<name>/<my_id>", middlewares=[middleware_one, early_return_middleware])
-    def get_lambda(my_id: str, name: str, injected: str) -> Response:
-        assert name == "my"
-        assert injected == "injected_value"
-
-        return Response(200, content_types.TEXT_HTML, my_id)
-
-    # WHEN calling the event handler
-    result = app(LOAD_GW_EVENT, {})
-
-    # THEN
-    assert result["statusCode"] == 400
-    assert result["multiValueHeaders"]["Content-Type"] == [content_types.TEXT_HTML]
-    assert result["body"] == "bad_response"
